// src/dto/estate/updateEstate.dto.ts
export class UpdateEstateDto {
  name: string;

  category: string;

  price: number;

  bed: number;

  toilet: number;

  m2: number;

  pool: boolean;

  imgUrl: string[];

  widthStreet: string;

  notification: string;

  otoStreet: boolean;

  gerion: string;

  status: number;

  type: string;

  userId: number;
}